Code cleanup and removal of 1.3 code.
