--- shared/src/main/scala/com/outr/reactify/DepConnector.scala
@@ -1,6 +1,0 @@
-package com.outr.reactify
-
-trait DepConnector[T, V] {
-  def combine(variable: => V, adjustment: => T): T
-  def extract(value: => T, adjustment: => T): V
-}
